%[ER: 105]% Made DH_PL and TH_PL run fine

+++ trunk/CEP/Transport/src/DH_PL.h
@@ -26,15 +26,29 @@
 #define LIBTRANSPORT_DH_PL_H
 
 #include <libTransport/DataHolder.h>		// for super-class definition
-#include <Common/LofarTypes.h>			// for ulong
+#include <Common/LofarTypes.h>			// for int64
 #include <PL/PLfwd.h>
 
 
 namespace LOFAR {
 
-/// DH_PL is a DataHolder implementation based on (LOFAR/Common) PL.  Note
-/// that the (current) implementation of PL uses DTL above ODBC above
-/// Postgresql.
+/// DH_PL is a DataHolder implementation based on (LOFAR/Common) PL.
+// Note that the (current) implementation of PL uses DTL which uses ODBC.
+// Usually a PostgreSQL database is used, but every ODBC-compliant database
+// system can be used.
+//
+// If a programmer wants it to make possible that a DataHolder can be stored
+// in a database, the particular DH class should be derived from DH_PL.
+// The data are stored in a blob in the database, and no fields are stored
+// individually, thus are not easily visible when accessing the database
+// directly. In that case the derived DH class does not need extra code.
+// An example for this case can be found in test/DH_Example.
+// <br>If one or more fields should be stored individually, the user needs
+// to add some extra code to tell PL which fields should be stored where.
+// An example can be found in test/DH_Example2. Please note that such
+// fields are also part of the blob, thus are stored twice. The only
+// advantage of storing them individually is that they are easier visible
+// in the database.
 
 class DH_PL : public DataHolder {
 
@@ -52,10 +66,10 @@ public:
   virtual PL::PersistentObject& getPO() const;		       
 
   // Pass the seqnr and get a reference to the PersistentObject.
-  PL::PersistentObject& preparePO (int id, int tag, int seqnr);		       
+  PL::PersistentObject& preparePO (int id, int tag, int64 seqnr);
   
   // Create a TPO object and set the table name in it.
-  // It should be overidden in a derived class having its own TPO object.
+  // It should be overridden in a derived class having its own TPO object.
   virtual void initPO (const string& tableName);
 
   // Get the variable values.
@@ -64,7 +78,7 @@ public:
     { return itsAppId; }
   int getTag() const
     { return itsTag; }
-  int getSeqNr() const
+  int64 getSeqNr() const
     { return itsSeqNr; }
   // </group>
 
@@ -76,14 +90,14 @@ private:
   // Forbid assignment.
   DH_PL& operator= (const DH_PL&);
   
-  int itsAppId;
-  int itsTag;
-  int itsSeqNr;
+  int   itsAppId;
+  int   itsTag;
+  int64 itsSeqNr;
   PO_DH_PL* itsPODHPL;
 };
  
  
-inline PL::PersistentObject& DH_PL::preparePO (int id, int tag, int seqnr)
+inline PL::PersistentObject& DH_PL::preparePO (int id, int tag, int64 seqnr)
 {
   itsAppId = id;
   itsTag   = tag;
